rag_system/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pydantic>=2.0.0
